services/ocr/requirements.txt
fastapi
uvicorn
python-multipart
# Stable CPU versions
paddlepaddle==2.6.2
paddleocr==2.7.3
# Numpy compatibility fix
numpy<2.0.0
# Headless OpenCV (Crucial for Docker)
opencv-python-headless
Pillow